+++ mydynalearn/drawer/visdom_drawer/visdom_batch_drawer/visdom_batch_drawer_comp_UAU.py
@@ -49,7 +49,7 @@ class VisdomBatchDrawerCompUAU(VisdomBatchDrawer):
                             A2_A2_true_pred,
                             A2_U_true_pred]
         self.performance_data_null_filtering(performance_data)
-        performance_data_type = [((index+1)*np.ones(data.shape[0])).astype(np.int)
+        performance_data_type = [((index+1)*np.ones(data.shape[0])).astype(np.int16)
                                  for index,data in enumerate(performance_data)]
         legend = ["U to U",
             "U to A1",
